+++ src/types.ts
@@ -6,6 +6,14 @@ export interface Product {
   discounts: Discount[];
 }
 
+export interface ProductForm {
+  name: string;
+  price: number;
+  stock: number;
+  description: string;
+  discounts: Discount[];
+}
+
 export interface Discount {
   quantity: number;
   rate: number;
@@ -19,6 +27,6 @@ export interface CartItem {
 export interface Coupon {
   name: string;
   code: string;
-  discountType: 'amount' | 'percentage';
+  discountType: "amount" | "percentage";
   discountValue: number;
 }
